+++ resources/js/admin.js
@@ -1,49 +1,51 @@
 document.addEventListener('DOMContentLoaded', function() {
     document.querySelectorAll('.role-checkbox').forEach((checkbox) => {
-        checkbox.addEventListener('change', function () {
+        checkbox.addEventListener('change', async function () {
             const userId = this.dataset.userId;
             const role = this.dataset.role;
             const isChecked = this.checked;
 
-            fetch(window.appRoutes.updateRole, {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                    'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]').content
-                },
-                body: JSON.stringify({
-                    user_id: userId,
-                    role: role,
-                    value: isChecked
-                })
-            })
-            .then(response => response.json())
-            .then(data => {
+            try {
+                const response = await fetch(window.appRoutes.updateRole, {
+                    method: 'POST',
+                    headers: {
+                        'Content-Type': 'application/json',
+                        'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]').content
+                    },
+                    body: JSON.stringify({
+                        user_id: userId,
+                        role: role,
+                        value: isChecked
+                    })
+                });
+
+                const data = await response.json();
+
                 if (data.success) {
                     if (userId === window.currentUserId) {
-                        updateSidebar();
+                        await updateSidebar();
                     }
                 } else {
                     console.error('Failed to update role');
                 }
-            })
-            .catch(error => {
+            } catch (error) {
                 console.error('There was a problem:', error);
-            });
+            }
         });
     });
 
-    function updateSidebar() {
-        const currentPath = window.location.pathname;
-        const currentRoute = currentPath.split('/').pop() || 'dashboard';
+    async function updateSidebar() {
+        try {
+            const currentPath = window.location.pathname;
+            const currentRoute = currentPath.split('/').pop() || 'dashboard';
 
-        fetch(window.appRoutes.sidebarRefresh, {
-            headers: {
-                'X-Requested-With': 'XMLHttpRequest'
-            }
-        })
-        .then(response => response.text())
-        .then(html => {
+            const response = await fetch(window.appRoutes.sidebarRefresh, {
+                headers: {
+                    'X-Requested-With': 'XMLHttpRequest'
+                }
+            });
+
+            const html = await response.text();
             const sidebarContainer = document.querySelector('nav.bg-gray-800');
             sidebarContainer.outerHTML = html;
 
@@ -58,9 +60,8 @@ document.addEventListener('DOMContentLoaded', function() {
                     link.classList.remove('hover:bg-gray-600');
                 }
             });
-        })
-        .catch(error => {
+        } catch (error) {
             console.error('Failed to update sidebar:', error);
-        });
+        }
     }
 });
